contract call trait - deprecated methods
